requirements.txt
wakeonlan==2.1.0
disnake==2.7.0
python-dotenv==0.21.0